+++ src/main/java/com/friendmonitor/models/activityupdate/BossKillCount.java
@@ -5,7 +5,7 @@ public class BossKillCount extends ActivityUpdate {
     private int npcId;
     private int count;
 
-    public BossKillCount(int npcId, int count, int accountHash) {
+    public BossKillCount(int npcId, int count, long accountHash) {
         super(accountHash, ActivityUpdateType.BOSS_KILL_COUNT);
         this.npcId = npcId;
         this.count = count;
