+++ src/main/java/com/friendmonitor/models/activityupdate/ItemDrop.java
@@ -3,7 +3,7 @@ package com.friendmonitor.models.activityupdate;
 public class ItemDrop extends ActivityUpdate {
     int itemId;
 
-    public ItemDrop(int itemId, int accountHash) {
+    public ItemDrop(int itemId, long accountHash) {
         super(accountHash, ActivityUpdateType.ITEM_DROP);
         this.itemId = itemId;
     }
