+++ src/main/java/com/friendmonitor/models/activityupdate/LevelUp.java
@@ -7,7 +7,7 @@ public class LevelUp extends ActivityUpdate {
     private Skill skill;
     private int level;
 
-    public LevelUp(Skill skill, int level, int accountHash) {
+    public LevelUp(Skill skill, int level, long accountHash) {
         super(accountHash, ActivityUpdateType.LEVEL_UP);
         this.skill = skill;
         this.level = level;
